Make The AbstractNamedBean setProperties method fire a property change listener.
