+++ java/src/jmri/implementation/AbstractIdTag.java
@@ -87,28 +87,5 @@ public abstract class AbstractIdTag extends AbstractNamedBean implements IdTag,R
         return Bundle.getMessage("BeanNameReporter");
     }
 
-    /**
-     * {@inheritDoc}
-     *
-     * This implementation fires property change listeners
-     * when properties change.
-     */
-    @Override
-    public void setProperty(String key,Object value){
-         Set<String> keySet = getPropertyKeys();
-         if(keySet.contains(key)){
-            // key already in the map, replace the value.
-            Object oldValue = getProperty(key);
-            if(!(oldValue.equals(value))){
-	       removeProperty(key); // make sure the old value is removed.
-	       super.setProperty(key,value); // then add the new one.
-               firePropertyChange(key,oldValue,value);
-            }
-         } else {
-	    super.setProperty(key,value);
-            firePropertyChange(key,null,value);
-         }
-    }
-
 //    private static final Logger log = LoggerFactory.getLogger(AbstractIdTag.class);
 }
